requirements.txt
numpy>=1.24,<2.0
faiss-cpu>=1.7.4
pyacoustid>=1.3.0
